+++ dark/dna.py
@@ -31,7 +31,6 @@ AMBIGUOUS = {
 BASES_TO_AMBIGUOUS = dict(
     (''.join(sorted(bases)), symbol) for symbol, bases in AMBIGUOUS.items())
 
-
 def compareDNAReads(read1, read2, matchAmbiguous=True, gapChars=('-'),
                     offsets=None):
     """
@@ -49,9 +48,10 @@ def compareDNAReads(read1, read2, matchAmbiguous=True, gapChars=('-'),
     @param read1: A C{Read} instance or an instance of one of its subclasses.
     @param read2: A C{Read} instance or an instance of one of its subclasses.
     @param matchAmbiguous: If C{True}, count ambiguous nucleotides that are
-        possibly correct as actually being correct. Otherwise, we are strict
-        and insist that only non-ambiguous nucleotides can contribute to the
-        matching nucleotide count.
+        possibly correct as actually being correct, and score these in the 
+        ambiguousMatchCount. Otherwise, we are strict and insist that only 
+        non-ambiguous nucleotides can contribute to the matching nucleotide 
+        count.
     @param gapChars: An iterable containing characters that should be
         considered to be gaps.
     @param offsets: If not C{None}, a C{set} of offsets of interest. Offsets
@@ -68,12 +68,29 @@ def compareDNAReads(read1, read2, matchAmbiguous=True, gapChars=('-'),
     read2AmbiguousOffsets = []
     empty = set()
 
+    def _identicalMatch(a, b):
+        return a == b and len(AMBIGUOUS[a]) == 1
+
+    def _ambiguousMatch(a, b, matchAmbiguous):
+        """
+        Checks if two characters match ambiguously if matchAmbiguous is True. 
+        A match is an ambiguous match if it is not an identical match, but the 
+        sets of ambiguous characters overlap.
+        """
+        return (matchAmbiguous and 
+                not _identicalMatch(a, b) and
+                AMBIGUOUS.get(a, empty) & AMBIGUOUS.get(b, empty))
+
     for offset, (a, b) in enumerate(zip_longest(read1.sequence.upper(),
                                                 read2.sequence.upper())):
         # Use 'is not None' in the following to allow an empty offsets set
         # to be passed.
         if offsets is not None and offset not in offsets:
             continue
+        if len(AMBIGUOUS.get(a, '')) > 1:
+            read1AmbiguousOffsets.append(offset)
+        if len(AMBIGUOUS.get(b, '')) > 1:
+            read2AmbiguousOffsets.append(offset)
         if a is None:
             # b has an extra character at its end (it cannot be None).
             assert b is not None
@@ -99,20 +116,15 @@ def compareDNAReads(read1, read2, matchAmbiguous=True, gapChars=('-'),
                     # a is a gap, b is not.
                     gapMismatchCount += 1
             else:
-                if len(AMBIGUOUS[a]) > 1:
-                    read1AmbiguousOffsets.append(offset)
                 if b in gapChars:
                     # b is a gap, a is not.
                     gapMismatchCount += 1
                     read2GapOffsets.append(offset)
                 else:
                     # Neither is a gap character.
-                    if len(AMBIGUOUS[b]) > 1:
-                        read2AmbiguousOffsets.append(offset)
-                    if a == b:
+                    if _identicalMatch(a, b):
                         identicalMatchCount += 1
-                    elif matchAmbiguous and (
-                            AMBIGUOUS.get(a, empty) & AMBIGUOUS.get(b, empty)):
+                    elif _ambiguousMatch(a, b, matchAmbiguous):
                         ambiguousMatchCount += 1
                     else:
                         nonGapMismatchCount += 1