+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.26 June 26, 2018
+
+Made a change in `dark/dna.py`, to make `identicalMatchCount` only count non-
+ambiguous matches. Added testfunction `testMatchwithIdenticalAmbuguity`.
+
 ## 3.0.25 June 26, 2018
 
 Added TravisCI Slack notifications.